+++ src/components/footer.js
@@ -1,0 +1,4 @@
+const Footer = () => `
+  <div class="footer"></div>
+`
+export default Footer
