+++ src/components/layout.js
@@ -1,0 +1,4 @@
+const Layout = (...children) => `
+  <div class="layout">${children.join('')}</>
+`
+export default Layout
