+++ src/components/note_text.js
@@ -1,0 +1,5 @@
+const NoteText = (note) => `
+  <h1>${note.title}</h1>
+  <p class="note-text">${note.text}</p>
+`
+export default NoteText
