+++ src/components/top_menu.js
@@ -1,0 +1,4 @@
+const TopMenu = () => `
+  <div class="top_menu"></div>
+`
+export default TopMenu
