+++ src/components/workspace.js
@@ -1,0 +1,4 @@
+const Workspace = (...children) => `
+  <div class="workspace">${children.join('')}</div>
+`
+export default Workspace
